Refactor metrics providers
